+++ plugit/api.py
@@ -39,6 +39,10 @@ class PlugItAPI(object):
             return u
         return None
 
+    def get_user_techgroups(self, userPk):
+        """Return the list of techgroups for a user"""
+        return self._request('user/{}/techgroups/'.format(userPk)).json()
+
     def get_subscription_labels(self, userPk):
         """Returns a list with all the labels the user is subscribed to"""
         r = self._request('subscriptions/' + str(userPk))
@@ -93,6 +97,10 @@ class PlugItAPI(object):
             retour.append(u)
 
         return retour
+
+    def get_techgroups(self):
+        """Return the list of techgroups"""
+        return self._request('techgroups/').json()
 
     def send_mail(self, sender, subject, recipients, message, response_id=None, html_message=False):
         """Send an email using EBUio features. If response_id is set, replies will be send back to the PlugIt server."""