+++ plugit_proxy/views.py
@@ -1172,6 +1172,39 @@ def api_get_project_members(request, key=None, hproPk=True):
     return HttpResponse(json.dumps({'members': liste}), content_type="application/json")
 
 
+def api_techgroup_list(request, key, hproPk):
+    """Return the list of techgroup"""
+
+    if not check_api_key(request, key, hproPk):
+        return HttpResponseForbidden
+
+    from users.models import TechGroup
+
+    retour = [{
+        'uuid': t.uuid,
+        'uid': t.uid,
+        'name': t.name,
+    } for t in TechGroup.objects.filter(is_enabled=True)]
+
+    return HttpResponse(json.dumps(retour), content_type="application/json")
+
+
+def api_user_techgroup_list(request, userPk, key, hproPk):
+    """Return the list of techgroup of a user"""
+
+    if not check_api_key(request, key, hproPk):
+        return HttpResponseForbidden
+
+    # From UUID to Pk
+    from users.models import TechUser
+
+    user = get_object_or_404(TechUser, pk=userPk)
+
+    retour = [t.uuid for t in user.techgroup_set.filter(is_enabled=True)]
+
+    return HttpResponse(json.dumps(retour), content_type="application/json")
+
+
 def generic_send_mail(sender, dests, subject, message, key, origin='', html_message=False):
     """Generic mail sending function"""